I've pulled out the functionality of the analyzer into a single python file which doesn't require all of the irrelevant config parameters (which would cause problems for external users).  I'll release this and the simple config file to 1KG for use in analyzing recalibration efforts.
Please note that this is literally my first foray into the wonderful world of python.  There could very well be a much more elegant way of releasing the script to external users without having to duplicate the file.  If so, anyone out there should (please) feel free to do so in a second release; but, for now, this needs to be online by tomorrow morning.


git-svn-id: file:///humgen/gsa-scr1/gsa-engineering/svn_contents/trunk@1404 348d0f76-0448-11de-a6fe-93d51630548a