I got bored today and decided to write the confusion matrix calculator. At present it is untested. I'm submitting it to subversion to make sure I have  previous revision to revert back to.
This is a calculator that will calculate:

P[ True base is X | read base mismatches, secondary base is Y, previous K bases are Z1,Z2,...ZK ]

where the number of pervious reference bases to take into account is user-defined. The secondary base is optional as well.

--usePreviousBases k

tells the walker to use the k previous reference bases in the transition table

--useSecondaryBase

tells the walker to use the secondary base at a locus in the transition table

these can be used together.



git-svn-id: file:///humgen/gsa-scr1/gsa-engineering/svn_contents/trunk@1816 348d0f76-0448-11de-a6fe-93d51630548a