Refactored read clipping framework into a generic utilities class, independent of ClipReadsWalker, which now uses this framework.  Some more cleanup is really needed, as some of the arguments to the classes are really only useful for ClipReads
ReduceReadsWalker -- does consensus-based read compression, v2.  Does all of the consensus calculations within the ConsensusReadCompressor per sample, and multi-sample case is handled by MultiSampleConsensusReadCompressor.  For deeply covered data sets, this projects a significant reduction in the number of mapped reads.  Impact on analysis call quality tbd.  Expected to be relatively minor, as the system automatically detects regions without a strong consensus, and expands a window around these so that +/- 10bp of all reads are shown around the unclear sites.  Not usable yet -- as it does not yet support streaming output, and actually holds all reads in memory at once.

git-svn-id: file:///humgen/gsa-scr1/gsa-engineering/svn_contents/trunk@5610 348d0f76-0448-11de-a6fe-93d51630548a